# api.yaml
openapi: 3.0.3
info:
  title: User Auth and Blog Api
  description: This the Documentation of of User Authentication System
  contact:
    email: apiteam@swagger.io
  license:
    name: Nodejs 18.16.0
    url: https://nodejs.org/en
  version: 1.0.11
servers:
  - url: http://localhost:5000
...
paths:
...
  /id:
    patch:
      tags:
        - Blog
      summary: update the existing blog by id
      description: you can edit blog partially
      operationId: updated
      requestBody:
        description: Write  a blog with title body and author
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/blog"
        required: true
      responses:
        "200":
          description: successful operation
      security:
        - user_auth:
            - write:blogs
            - read:blogs
        - api_key: []
    delete:
      tags:
        - Blog
      summary: Delete the blog by id
      description: get blog by id and delete  it
      operationId: Deleted
      responses:
        "400":
          description: Invalid ID supplied
        "404":
          description: Order not found
      security:
        - user_auth:
            - write:blogs
            - read:blogs
        - api_key: []
components:
  schemas:
    blog:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 10
        userId:
          type: integer
          format: int64
          example: 12
        author:
          type: string
          example: Umar/Ali/Saad
        title:
          type: string
          example: technology
        body:
          type: string
          description: Order Status
          example: write the blog body here
...
  securitySchemes:
    user_auth:
      type: oauth2
      flows:
        implicit:
          authorizationUrl: http://localhost:5000/oauth/authorize
          scopes:
            write:pets: modify pets in your account
            read:pets: read your pets
    api_key:
      type: apiKey
      name: api_key
      in: header